+++ EssentialFeed/EssentialFeediOSTests/Controllers/FeedViewControllerTests.swift
@@ -280,60 +280,4 @@ final class FeedViewControllerTests: XCTestCase {
                          url: URL = URL(string: "http://any-url.com")!) -> FeedImage {
     return FeedImage(id: UUID(), description: description, location: location, url: url)
   }
-  
-  class LoaderSpy: FeedLoader, FeedImageDataLoader {
-   
-    // MARK: - FeedLoader
-    
-    private var feedRequests = [(FeedLoader.Result) -> Void]()
-    
-    var loadFeedCallCount: Int {
-      return feedRequests.count
-    }
-    
-    func load(completion: @escaping (FeedLoader.Result) -> Void) {
-      feedRequests.append(completion)
-    }
-    
-    func completFeedLoading(with feed: [FeedImage] = [], at index: Int) {
-      feedRequests[index](.success(feed))
-    }
-    
-    func completFeedLoadingWithError(at index: Int) {
-      let error = NSError(domain: "an error", code: 0)
-      feedRequests[index](.failure(error))
-    }
-    
-    // MARK: - FeedImageDataLoader
-    
-    private struct TaskSpy: FeedImageDataLoaderTask {
-      let cancelCallback: () -> Void
-      func cancel() {
-        cancelCallback()
-      }
-    }
-    
-    private var imageRequests = [(url: URL, completion: (FeedImageDataLoader.Result) -> Void)]()
-    
-    var loadedImageURLs: [URL] {
-      return imageRequests.map { $0.url }
-    }
-    
-    private(set) var cancelledImageURLs = [URL]()
-    
-    func loadImageData(from url: URL,
-                       completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
-      imageRequests.append((url, completion))
-      return TaskSpy { [weak self] in self?.cancelledImageURLs.append(url) }
-    }
-  
-    func completeImageLoading(with imageData: Data = Data(), at index: Int = 0) {
-      imageRequests[index].completion(.success(imageData))
-    }
-    
-    func completeImageLoadingWithError(at index: Int = 0) {
-      let error = NSError(domain: "an error", code: 0)
-      imageRequests[index].completion(.failure(error))
-    }
-  }
 }
